+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Don't scroll input editor when rewinding (#381)
 * Raise the limit of thread number in states
 * State loading doesn't write zeros on zero pages, preventing allocations
+* Switch input mapping to tabs
 
 ### Fixed